requirements.txt
tensorflow
keras
opencv-python
pillow
numpy
pandas
matplotlib
scikit-learn
albumentations
seaborn
streamlit
tqdm
